requirements.txt
torch
torchvision
PyYAML
diffusers
transformers
typing
deepspeed
numpy
pillow
scikit-learn
accelerate
wandb